refactor(logdrains): rename log_drains to logdrains
